requirements_rag_cag.txt
# RAG-CAG Analytics Dependencies
# Install with: pip install -r requirements_rag_cag.txt

# Database connections
psycopg2-binary==2.9.7
pyodbc==4.0.39

# RAG and embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3

# Data processing
pandas==2.0.3
pyyaml==6.0.1

# Logging and utilities
python-dotenv==1.0.0
click==8.1.7

# Optional: Vector database support
# chromadb==0.4.0  # Uncomment for production vector store
# pinecone-client==2.2.4  # Uncomment for Pinecone integration